feat: 评论量

+++ springboot/system/src/main/java/org/example/springboot/system/domain/entity/Comment.java
@@ -43,7 +43,7 @@ public class Comment extends BaseEntity {
      * 业务类型
      */
     @Schema(description = "业务类型")
-    private Byte bizType;
+    private Integer bizType;
     /**
      * 内容
      */
